requirements.txt
Django==3.2.9
django_debug_toolbar==3.7.0
django_filter==21.1
djangorestframework==3.14.0
django-debug-toolbar-force==0.1.8
gunicorn==20.1.0
psycopg2-binary==2.9.5